Mega Codegen Commit
This commit is large. This is one of those chunks of work that I wasn't
100% sure how was going to work or look until I actually did it. So
that's my excuse.

The meat of this commit is codegen/lco/ which contains a script and a
template to generate LCO instrument definitions from the OCS instruments
info API. The result is then checked in as
src/aeonlib/ocs/lco/instruments.py where there is a Pydantic model per
instrument for use in construction of observation requests.

The generator code will not be included in the final built module.