Proper serialiation of the optical elements field.
This is a little complex but worth it to clean up the API in my
opinion. We save ourselves the need to generate a bunch of
per-instrument classes to hold on to the optical elements available for
each instrument (this is how the remote API should work too, but alas).