Module-level Window model
Moves the Window model out of the OCS specific models.py into a top
level models.py making it a more logical import for other facilities.

The ESO AbsoluteTimeContraint model now as a `from_window` constructor
that accepts a Window model, meaning a single Window model that is
defined in a script can be used for both facilties.