Custom Angle Type and tests
Adds a custom Angle type that is a union of astropy.coordinates.Angle
and str, meaning that models that specify it as the type for a field can
be passed either the atropy Angle types or strings. Strings will attempt
to be coverted to Angles internally using astropy Angle parsing logic.
When these models are serialized the string representation of the angle
will be provided, such as in JSON.

Additionally a small modification to the Time type was made to indicate
to type checkers that datetimes are accepted as well as Time objects
without having to explicitly set the union on models that use it.