Add SOAR specific AEONlib module
Requests to SOAR use the LCO infrastructure, so this new facility is a
thin wrapper around the LcoFacility with the only real change being the
ability to use explicit SOAR configuration. This builds on the previous
commit of splitting the instruments into SOAR/NON-SOAR modules making
this SOAR module as a whole function as a standalone entity.