Update documentation for NonSiderealTarget
The two fields at the end are needed for major planets only. The mean
longitude and longitude of perihelion are related to the argument of
perihelion and mean anomaly (needed for asteroids); details are shown in
step 2. of 'Formulae for using the Keplerian elements' on https://ssd.jpl.nasa.gov/planets/approx_pos.html